chore: Add a constructor to OpaqueHandle and move Nlm4Lock to lock.rs
